Assignment Task1
This is the repository to task 1, with the first two subparts.
archibus_1 : the original file upload
src_uploadtoFS: is the first subpart where the file is uploaded from
the file system, this is just the src folder, which when replaced in
archibus_1 should just work fine.

Spring4MVCFileDownloadExample: is the part 2 of task one, done using
servlets